test(ct): rebase ct.tests on current master, move tests that require Internet access into separate module, move selenium tests into bdd module
By default ``functional`` and ``bdd`` tests are not started.
Also fix some tests according to new code in master.
