+++ mysite/ct/tests/bdd/__init__.py
@@ -1,0 +1,1 @@
+from selenium_integrate import *
