+++ mysite/ct/tests/functional/__init__.py
@@ -1,0 +1,1 @@
+from integrate import *
